hls: add p_tcp_consumer() process to consume the incoming TCP payloads, and trigger some collections when receiving payloads on session #64
This is purely example code to show how to integrate the TCP logic in the design and should be modified by our customers.

Session #64 is not available for now.
TB (testbench) has been improved to allow testing of TCP input streams.

We had to workaround the Vivado HLS code generation by using input data in the output.
Otherwise, the generated code does not compile.