Improve TCP consumer process
- TCP: Add tcp error bit check in consumer
- TCP: Add DMA output to TCP consumer

NOTE: Adding a DMA output to the TCP consumer process causes the cosim
to fail. A service request was filed to Xilinx, which replied that the
cosim feature is not supported for designs with top-level interfaces
relying on ``ap_ctrl_none`` flow control (which is the case of the
current implementation).