Fix test-suite disabled collection check
This commit changes the `standalone_trigger.test_standalone_trigger_disabled_collection.StandAloneTriggerDisabledCollectionTest` by expecting arguments of the collection which is armed.

Previously the expected arguments where the ones from the first
triggered collection which is disarmed. Following integration of the
feature #109983, these arguments are discarded along with the
collection, meaning the relevant test had to be updated (hence this
present commit).

refs #110458